// src/types/user-panel/my-courses-list.ts
export interface MyCoursesList {
  teacherName: string;
  fullName: string;
  termName: string;
  statusName: string;
  levelName: string;
  courseTitle: string;
  courseId: string;
  describe: string;
  cost: string;
  lastUpdate: string;
  tumbImageAddress: string | null;
  classRoomName: string;
  typeName: string;
  isExpire: boolean;
  isActive: boolean;
  isdelete: boolean;
  teacherId: number;
  groupName: string;
  courseGroupId: number;
  studentId: number;
  paymentStatus: string;
}